refactor: order 상태변화 payment 응답 처리 하는중(KAN-21)

+++ client/order/src/main/java/com/live_commerce/order/infrastructure/client/PaymentReadyRequestDto.java
@@ -5,6 +5,6 @@ import java.util.UUID;
 
 public record PaymentReadyRequestDto(
         UUID orderId,
-        double amount,
+        BigDecimal amount,
         String itemName
 ) {}
